% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclicDescent.R
\name{cyclicDescent}
\alias{cyclicDescent}
\title{Cyclic Descent}
\usage{
cyclicDescent(
  x,
  t = NULL,
  trend = FALSE,
  ip = NULL,
  lp = NULL,
  step = NULL,
  hn = NULL,
  neig = 0,
  exclude = NULL,
  alpha = 0.05,
  rrss = FALSE,
  plots = FALSE
)
}
\arguments{
\item{x}{Vector of values to be analyzed.}

\item{t}{Vector of time (optional). If specified, it has to have the same length as \code{x} and must be monotonically increasing.}

\item{trend}{Logical. If \code{TRUE}, the linear tendency is calculated and subtracted from the data.}

\item{ip}{Initial period to test.}

\item{lp}{Final period to test.}

\item{step}{Is the increment between periods. Defaults to 1, assuming that two successive data values are separated by one unit of time.}

\item{hn}{Number of harmonics to estimate. If not specified, the function will stop when the increase in R-squared is not significative or when a maximum of 30 harmonics is reached. This behaviour is overrid when a value is provided.}

\item{neig}{By default, once an optimal period (op) is identified, it is excluded from subsequent searches, unless neig = −1. Positive integers will also exclude neighboring periods (op − neig and op + neig).}

\item{exclude}{The periods given here will be excluded in the periods search.}

\item{alpha}{The confidence level for the statistical test in the cyclic descent. Defaults to 0.05.}

\item{rrss}{Logical. When \code{TRUE}, the reciprocal of the residual sum of squares in every step of the cyclic descent will be returned.}

\item{plots}{Logical, when TRUE, the RRSS versus the tested periods and the cumulated harmonics fit in every step of the cyclic descent will be plotted.}
}
\value{
A list with two elements, (1) 'harmonics' a six-column data frame giving the periods, amplitudes, phases, lags, RSS and R-squared. Note that after the first model, the following will include the new period referred to and those already found. (2) 'Stats' this data frame provides the F statistics, with the corresponding degrees of freedom and associated p-values for every model comparisons.
}
\description{
This function implements the cyclic descent with periodoc regression to find periodicities in a regular time series.
}
\details{
This function can ...
}
\examples{
# load simulated data
data(sim)
# find periodicites
sim.cd <- cyclicDescent(x = sim)
sim.cd

}
\references{
González-Rodríguez E, Villalobos H, Gómez-Muñoz VM and Ramos-Rodríguez A. 2015. Computational Method for Extracting and Modeling Periodicities in Time Series. Open Journal of Statistics 05(06): 604–17. \url{https://doi.org/10.4236/ojs.2015.56062}.
}
\seealso{
\code{\link{lm.harmonics}}.
}
\author{
Héctor Villalobos.
}
