% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHarmonics.R
\name{harmonics}
\alias{harmonics}
\title{harmonicos}
\usage{
harmonics(fit.lm)
}
\arguments{
\item{fit.lm}{}
}
\value{
objeto de clase 'lm'
}
\description{
Read a specific variable from several global CMEMS files, for choosen area 
limits (and possibly a certain depth).
}
\details{
This function can ...
}
\examples{
data(sim)
# descenso cíclico
sim.cd <- cyclicDescent(x = sim)
# periodos significativos
p <- sim.cd$harmonics$Period[1:4]
sim.fit <- lm.harmonics(x = sim, periods = p, trend=FALSE)
harmonics(sim.fit)

}
\seealso{
\code{\link{cyclicDescent}}.
}
\author{
Héctor Villalobos.
}
