% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHarmonics.R
\name{makeHarmonics}
\alias{makeHarmonics}
\title{harmonicos}
\usage{
makeHarmonics(fit.lm, plot = TRUE)
}
\arguments{
\item{fit.lm}{}

\item{plot.}{}
}
\value{
objeto de clase 'lm'
}
\description{
Read a specific variable from several global CMEMS files, for choosen area 
limits (and possibly a certain depth).
}
\details{
This function can ...
}
\seealso{
\code{\link{cyclicDescent}}.
}
\author{
Héctor Villalobos.
}
