% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicRegModel.R
\name{periodicRegModel}
\alias{periodicRegModel}
\title{Periodic Regression Model}
\usage{
periodicRegModel(x, t = NULL, periods, center.x = TRUE, trend = FALSE)
}
\arguments{
\item{x}{Vector containing the time series to be fitted.}

\item{t}{Vector of time. Must be monotonically increasing of the same length of x.}

\item{periods}{Vector of periods, typically found with \code{cyclicDescent}.}

\item{center.x}{Logical. If \code{TRUE}, the vector x is centered to mean zero.}

\item{trend}{Logical. If \code{TRUE}, the linear trend is included in the model.}
}
\value{
A list with two elements, a data frame with variables and the periodic regression model to be passed to \code{lm}.
}
\description{
Build the periodic regression model according to given periods.
}
\details{
This function can ...
}
\examples{
# load simulated data
data(sim)
# find periodicities
sim.cd <- cyclicDescent(x = sim)
# significant or desired periods
p <- sim.cd$harmonics$Period[1:4]
model to fit with lm
perReg <- periodicRegModel(x = sim, periods = p, center.x = FALSE, trend = FALSE)
perReg
# model fit
summary(lm(perReg$model, data = perReg$data))
}
\seealso{
\code{\link{cyclicDescent}}.
}
\author{
Héctor Villalobos.
}
