% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_periodicReg.R
\name{plot_periodicReg}
\alias{plot_periodicReg}
\title{Plot a Periodic Regression Model}
\usage{
plot_periodicReg(fit.lm)
}
\arguments{
\item{fit.lm}{}
}
\value{
objeto a plot is produced
}
\description{
Read a specific variable from several global CMEMS files, for choosen area 
limits (and possibly a certain depth).
}
\details{
This function can ...
}
\examples{
data(sim)
# descenso cíclico
sim.cd <- cyclicDescent(x = sim)
# periodos significativos
p <- sim.cd$harmonics$Period[1:4]
# build Model
perReg <- periodicRegModel(x = sim, periods = p, center.x = FALSE)
plot(perReg)

}
\seealso{
\code{\link{cyclicDescent}}.
}
\author{
Héctor Villalobos.
}
