% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclicDescent.R
\name{cyclicDescent}
\alias{cyclicDescent}
\title{Descenso cíclico}
\usage{
cyclicDescent(
  x,
  t = NULL,
  trend = FALSE,
  ip = NULL,
  lp = NULL,
  step = NULL,
  hn = NULL,
  neig = 0,
  exclude = NULL,
  alpha = 0.05,
  rrss = FALSE,
  plots = c("all", "last", "none")
)
}
\arguments{
\item{neig}{}

\item{x.}{}

\item{t.}{}

\item{trend.}{}

\item{ip.}{}

\item{lp.}{}

\item{step.}{}

\item{hn.}{}

\item{neig.}{}

\item{exclude.}{}

\item{alpha.}{}

\item{rrss.}{}

\item{plots.}{}
}
\value{
Una lista con dos elementos, 'harmonics' (data frame) con seis columnas 
(Periodo, amplitud, fase, lag, RSS, y R-squared)
}
\description{
Read a specific variable from several global CMEMS files, for choosen area 
limits (and possibly a certain depth).
}
\details{
This function can ...
}
\seealso{
\code{\link{lm.harmonics}}.
}
\author{
Héctor Villalobos.
}
