% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclicDescent.R
\name{cyclicDescent}
\alias{cyclicDescent}
\title{Cyclic Descent}
\usage{
cyclicDescent(
  x,
  t = NULL,
  trend = FALSE,
  ip = NULL,
  lp = NULL,
  step = NULL,
  hn = NULL,
  neig = 0,
  exclude = NULL,
  alpha = 0.05,
  rrss = FALSE,
  plots = FALSE
)
}
\arguments{
\item{x}{Vector containing the time series to be analyzed.}

\item{t}{Vector of time. Must be monotonically increasing of the same length of x.}

\item{trend}{Logical. If \code{TRUE}, the linear tendency is calculated and subtracted from the data.}

\item{ip}{Initial period to test. Must be in the same units as \code{t}.}

\item{lp}{Final period to test.}

\item{step}{Period step. Is the increment in time for each tested period according to the time unit (derived from t above), the default value is 1 assuming that two successive data values are separated by one unit of time.}

\item{hn}{Number of harmonics to estimate. Determined by the MRRSS criteria or an integer number supplied by the user.}

\item{neig}{By default, a period identified (op) in a periodic regression will not be considered in a subsequent search unless neig = −1. Other positive integers will cause the function to also exclude neighboring periods: op − neig and op + neig in subsequent searches, note that the neighbors are directly related to the step argument.}

\item{exclude}{The periods given here will be excluded in the periods search..}

\item{alpha}{The confidence level for the statistical test in the cyclic descent or probability error if the corresponding model fit is rejected. Defaults to alpha = 0.05.}

\item{rrss}{Reciprocal of residual sum of squares (string). Logical, defaults to FALSE. If TRUE, the periods tested and their corresponding RRSS in every step of the cyclic descent will be provided.}

\item{plots}{Logical, when TRUE, the RRSS versus the tested periods and the cumulated harmonics fit in every step of the cyclic descent will be plotted.}
}
\value{
Una lista con dos elementos, 'harmonics' (data frame) con seis columnas 
(Periodo, amplitud, fase, lag, RSS, y R-squared)
}
\description{
This function implements the cyclic descent with periodoc regression to find periodicities in a regular time series.
}
\details{
This function can ...
}
\examples{
# load simulated data
data(sim)
# find periodicites
sim.cd <- cyclicDescent(x = sim)
sim.cd

}
\seealso{
\code{\link{lm.harmonics}}.
}
\author{
Héctor Villalobos.
}
