% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.harmonics.R
\name{lm.harmonics}
\alias{lm.harmonics}
\title{multiple regression}
\usage{
lm.harmonics(x, t = NULL, periods, trend = FALSE)
}
\arguments{
\item{x.}{}

\item{t.}{}

\item{periods.}{}

\item{trend.}{}
}
\value{
objeto de clase 'lm'
}
\description{
Read a specific variable from several global CMEMS files, for choosen area 
limits (and possibly a certain depth).
}
\details{
This function can ...
}
\seealso{
\code{\link{cyclicDescent}}.
}
\author{
Héctor Villalobos.
}
